+++ src/components/Channel/Channel.tsx
@@ -125,6 +125,7 @@ type ChannelPropsForwardedToComponentContext<
   | 'LinkPreviewList'
   | 'LoadingIndicator'
   | 'Message'
+  | 'MessageActions'
   | 'MessageBouncePrompt'
   | 'MessageDeleted'
   | 'MessageListNotifications'
@@ -1241,6 +1242,7 @@ const ChannelInner = <
       LinkPreviewList: props.LinkPreviewList,
       LoadingIndicator: props.LoadingIndicator,
       Message: props.Message,
+      MessageActions: props.MessageActions,
       MessageBouncePrompt: props.MessageBouncePrompt,
       MessageDeleted: props.MessageDeleted,
       MessageListNotifications: props.MessageListNotifications,
@@ -1293,6 +1295,7 @@ const ChannelInner = <
       props.LinkPreviewList,
       props.LoadingIndicator,
       props.Message,
+      props.MessageActions,
       props.MessageBouncePrompt,
       props.MessageDeleted,
       props.MessageListNotifications,