+++ src/components/Message/MessageSimple.tsx
@@ -71,9 +71,12 @@ const MessageSimpleWithContext = <
     Attachment = DefaultAttachment,
     Avatar = DefaultAvatar,
     EditMessageInput = DefaultEditMessageForm,
+    MessageOptions = DefaultMessageOptions,
+    // TODO: remove this "passthrough" in the next
+    // major release and use the new default instead
+    MessageActions = MessageOptions,
     MessageDeleted = DefaultMessageDeleted,
     MessageBouncePrompt = DefaultMessageBouncePrompt,
-    MessageOptions = DefaultMessageOptions,
     MessageRepliesCountButton = DefaultMessageRepliesCountButton,
     MessageStatus = DefaultMessageStatus,
     MessageTimestamp = DefaultMessageTimestamp,
@@ -174,7 +177,7 @@ const MessageSimpleWithContext = <
             onClick={handleClick}
             onKeyUp={handleClick}
           >
-            <MessageOptions />
+            <MessageActions />
             <div className='str-chat__message-reactions-host'>
               {hasReactions && <ReactionsList reverse />}
             </div>